Move Veryl lowering diagnostics into frontend
